core: move BlockLocation to sim_infra.api
Allows larger use and exposition later
